refactors reducers
