+++ project/src/store/middlewares/redirect.ts
@@ -1,11 +1,11 @@
 import browserHistory from 'browser-history';
 import {Middleware} from 'redux';
-import {filmCatalogReducer} from 'store/reducers/film-catalog-reducer';
-import {currentFilmReducer} from 'store/reducers/current-film-reducer';
+import {catalogFilmsReducer} from 'store/catalog-films/catalog-films-reducer';
+import {currentFilmReducer} from 'store/current-film/current-film-reducer';
 import {ActionType} from 'types/action';
 
 type Reducer =
-  | ReturnType<typeof filmCatalogReducer>
+  | ReturnType<typeof catalogFilmsReducer>
   | ReturnType<typeof currentFilmReducer>;
 
 export const redirect: Middleware<unknown, Reducer> =
