+++ project/src/store/root-reducer.ts
@@ -1,14 +1,27 @@
 import {combineReducers} from 'redux';
-import {filmCatalogReducer} from 'store/reducers/film-catalog-reducer';
-import {promoFilmReducer} from 'store/reducers/promo-film-reducer';
-import {favoriteFilmsReducer} from 'store/reducers/favorite-films-reducer';
-import {currentFilmReducer} from 'store/reducers/current-film-reducer';
+import {catalogFilmsReducer} from 'store/catalog-films/catalog-films-reducer';
+import {promoFilmReducer} from 'store/promo-film/promo-film-reducer';
+import {favoriteFilmsReducer} from 'store/favorite-films/favorite-films-reducer';
+import {currentFilmReducer} from 'store/current-film/current-film-reducer';
+import {userAuthorizationReducer} from 'store/user-authorization/user-authorization-reducer';
+import {genresReducer} from 'store/genres/genres-reducer';
+
+export enum NameSpace {
+  catalog = 'CATALOG',
+  promo = 'PROMO',
+  favorite = 'FAVORITE',
+  current = 'CURRENT',
+  user = 'USER',
+  genres = 'GENRES',
+}
 
 export const rootReducer = combineReducers({
-  filmCatalog: filmCatalogReducer,
-  promoFilm: promoFilmReducer,
-  favoriteFilms: favoriteFilmsReducer,
-  currentFilm: currentFilmReducer,
+  [NameSpace.catalog]: catalogFilmsReducer,
+  [NameSpace.promo]: promoFilmReducer,
+  [NameSpace.favorite]: favoriteFilmsReducer,
+  [NameSpace.current]: currentFilmReducer,
+  [NameSpace.user]: userAuthorizationReducer,
+  [NameSpace.genres]: genresReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
