--- project/src/types/favorite-films-state.ts
@@ -1,5 +1,0 @@
-import {Films} from 'types/film';
-
-export type FavoriteFilmsState = {
-  favoriteFilms: Films;
-}
