--- project/src/types/promo-film-state.ts
@@ -1,5 +1,0 @@
-import {Film} from 'types/film';
-
-export type PromoFilmState = {
-  promoFilm: Film | null;
-}
