+++ project/src/types/state.ts
@@ -1,15 +1,38 @@
 import {AuthorizationStatus} from 'configs/auth-status';
-import {Films, GenreName} from 'types/film';
+import {Film, Films, GenreName} from 'types/film';
 import {UserInfo} from 'types/user-info';
+import {Comments} from 'types/comment';
 
-export type State = {
+export type UserAuthorization = {
+  authorizationStatus: AuthorizationStatus;
+  userInfo: UserInfo;
+  loginError: string;
+};
+
+export type Genres = {
   currentGenre: GenreName;
   genres: Array<GenreName>;
+}
+
+export type State = {
   films: Films;
   filteredFilms: Films;
   currentPage: number;
-  authorizationStatus: AuthorizationStatus;
   isDataLoaded: boolean;
-  userInfo: UserInfo;
-  loginError: string;
-}
+};
+
+export type FilmState = {
+  currentFilm: Film | null;
+  similarFilms: Films;
+  filmComments: Comments;
+  isCurrentFilmLoaded: boolean;
+  isCommentPosting: boolean;
+};
+
+export type FavoriteFilmsState = {
+  favoriteFilms: Films;
+};
+
+export type PromoFilmState = {
+  promoFilm: Film | null;
+};
